상수 분리 및 formatCurrency 함수 생성
